// src/model/dto/EmployeeDto.ts
export interface EmployeeDto {
  emp_id: string; // employee id
  project_id: string; // employee project id
  date_from: string; // employee start work date on project
  date_to: string; // employee end work date on project
  days_worked: number; // number of days worked together
  partner_id: string;  // 
}